+++ openbox/stacking.c
@@ -49,7 +49,7 @@ void stacking_set_list()
     }
 
     PROP_SETA32(RootWindow(ob_display, ob_screen),
-                net_client_list_stacking, window, (guint32*)windows, i);
+                net_client_list_stacking, window, (gulong*)windows, i);
 
     g_free(windows);
 }
